Moved base text fields.
Moved base text fields in order to add different custom widgets.
