+++ lib/view/widget/text/pictron_password_text_field.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:pictron/view/widget/pictron_base_text_field.dart';
+import 'package:pictron/view/widget/text/pictron_base_text_field.dart';
 
 class PictronPasswordTextField extends PictronBaseTextField {
   PictronPasswordTextField({
